replace "address" element on all page footers with "p" element
the reason for this change is the instructor's feedback that
"[...] address should not be used for the contact address".

the HTML5 specs say this about the "address" element:

  The address element represents the contact information for its
  nearest article or body element ancestor. If that is the body
  element, then the contact information applies to the document as
  a whole.
  [...]
  The address element must not be used to represent arbitrary addresses
  (e.g., postal addresses), unless those addresses are in fact the
  relevant contact information. (The p element is the appropriate
  element for marking up postal addresses in general.)

after reading this, and also other resources on the web, I
personally believe that I was correct in using "address" for the
postal address of the website owner.