menu button is now an "img" element
the previous solution used the :before pseudo-element to add a menu
button to the dropdown container. this worked fine in desktop browsers,
but didn't work at all in mobile browsers. touch events in mobile
browsers on the :before pseudo-element apparently don't trigger the
:hover selector of the dropdown container - at least that was the case
in Mobile Safari with which I tested.

the new solution replaces the :before pseudo-element with a real element
in the HTML markup. because the dropdown button displays an image, the
simplest solution is to use an "img" element.